REfactored a bit
